requirements.txt
PyQt5 == 5.15.10
tensorflow == 2.10
numpy == 1.26.4
keras == 3.1.1
matplotlib == 3.8.3
scipy == 1.13.0
pillow == 10.3.0
scikit-learn == 1.5.1